Add reference attribute to Wizard::Base
When we make requests to the API it is able to determine the
client/service making the request, however it would be beneficial
to include a transaction-level reference for certain metrics.

Add `reference` attribute to `Wizard::Base` that populates with
the name of the wizard/transaction. For example,
`MailingListWizard` will have the reference `mailing_list`.