Skip steps the candidate has already seen/submitted
At the moment if a candidate completes a wizard journey and then opts to
change one of their answers from a review page it results in them needing to
re-submit all the following steps to get back to the review page.

GDS recommend we instead take the user directly back to the review page. If a
change in their answers causes a new question to be asked we should still take
them to this question before going back to the review answers page. If the
candidate changes an answer and then qualifies for an exit step
(`#can_proceed?` is `false`) they should be taken to this instead of the review
answers page.