Ensure initial sign up request returns 301
The way the wizard is setup it provides two paths to the initial step of the
multi-stage form:

```
/sign_up
/sign_up/first_step
```

If we provide a link to the index path (`/sign_up`) then the user is redirected
with a 302 HTTP status code (temporary redirect) by the Rails `redirect_to`
helper. Instead, we should be using a 301 status code here for a permanent
redirect.